Trim trailing rows from the google sheet when uploading.
The batchUpdate(pasteData) method used to upload a CSV overwrites existing
lines, but does not strictly replace the entire sheet.  If the google sheet
already contains more lines than are being pasted from the CSV, any old lines
beyond the end of the CSV file will remain in the sheet.

If we are reducing the number of lines in the sheet then we need to clear those
trailing lines.  So use an updateCells method to clear the values of all rows
below the bottom of the CSV.

Using userEnteredValue with no value will clear the contents while leaving any
formatting intact.