Partial changeover of spreadsheet update API usage
Turn from pasteData/PASTE_VALUES into a values.update upload; rather than
pasting the raw CSV, we now read the CSV locally and upload as cell contents
directly.

It should have the same effect most of the time, except that pasteData mimics a
user ctrl-V paste, which is affected by the current filter view scope; so if
there is a filter view active, the upload is corrupted as we paste the full
sheet into a subset of the spreadsheet lines.

The values.update API is not constrained by the user view, so it correctly
updates the full sheet even if the user is only viewing a filtered subset.