Clean up file handling for 3-way merge
Make sure we leave the results of 3-way merge in the main local
file, not just in the temporary .MERGE file, for the user to edit
in place for conflict resolution.

We maintain an old copy of the local file now so can always restore from
that if we need to.